Configured the Pundit scope for better performance.
Removed the check in ApplicationController that redirected to the new_company_path if a company wasn't present.
That was generating a SQL query with every request!

Instead configured Devise and Pundit to handle things more efficiently.

Namespaced the controllers that need to be scoped through the company and added a scope to the route file.

Added a drop-down menu for the app nav to sign out.